#DOC-349 EXTRA restructured migration codebase, more console logging on operations
