--- migration-tools/content-structure.json
@@ -1,31 +1,0 @@
-{
-
-    "move": {
-        "3.10": {
-            "arangosync": "deployment",
-            "satellites": "deployment",
-            "smartjoins": "deployment"
-        },
-        "3.11": {
-            "arangosync": "deployment",
-            "satellites": "deployment",
-            "smartjoins": "deployment"
-        }
-    },
-    "restructure": {
-        "3.10": {
-            "introduction": ["about-arangodb", "getting-started", "modeling-data", "programs-tools"],
-            "indexing": ["analyzers"],
-            "deploying": ["architecture", "deployment"],
-            "operations": ["installation", "uninstallation", "backup-restore", "downgrading", "administration", "security" , "monitoring", "troubleshooting"],
-            "developer": ["http", "drivers", "foxx-microservices"] 
-        },
-        "3.11": {
-            "introduction": ["about-arangodb", "getting-started", "modeling-data", "programs-tools"],
-            "indexing": ["analyzers"],
-            "deploying": ["architecture", "deployment"],
-            "operations": ["installation", "uninstallation", "backup-restore", "downgrading", "administration", "security" , "monitoring", "troubleshooting"],
-            "developer": ["http", "drivers", "foxx-microservices"] 
-        }
-    }
-}
