--- migration-tools/links.py
@@ -1,3 +1,0 @@
-brokenLinks = {
-    "../aql/": "/core-topics/aql/"
-}
